Move LogSurgeonReader into its own file
