+++ components/core/src/ReaderInterface.cpp
@@ -117,14 +117,3 @@ size_t ReaderInterface::get_pos () {
 
     return pos;
 }
-
-ReaderInterfaceWrapper::ReaderInterfaceWrapper (ReaderInterface& reader_interface)
-        : m_reader_interface(reader_interface) {
-    read = [this] (char* buf, size_t count, size_t& read_to) -> log_surgeon::ErrorCode {
-        m_reader_interface.read(buf, count, read_to);
-        if (read_to == 0) {
-            return log_surgeon::ErrorCode::EndOfFile;
-        }
-        return log_surgeon::ErrorCode::Success;
-    };
-}
